+++ run_table8.py
@@ -65,20 +65,20 @@ for d in ds:
     for k in ks:
         b    = max(32, k)
         np.random.seed(42)
-        X_np = np.random.randn(N, m, d).astype(np.float64)
+        X_np = np.random.randn(N, m, d).astype(np.float32)
         X    = torch.from_numpy(X_np)
-        Xn   = torch.empty((N, m), dtype=torch.float64)
-        Dloc = torch.zeros((N, m, b), dtype=torch.float64)
+        Xn   = torch.empty((N, m), dtype=torch.float32)
+        Dloc = torch.zeros((N, m, b), dtype=torch.float32)
 
         if args.pipeline == "fused" and k == b:
             from knn_kokkos_keqb import run_knn_pipeline_keqb
-            Gdst = torch.full((N, m, 2 * k), torch.finfo(torch.float64).max, dtype=torch.float64)
+            Gdst = torch.full((N, m, 2 * k), torch.finfo(torch.float32).max, dtype=torch.float32)
             Gidx = torch.full((N, m, 2 * k), -1,                             dtype=torch.int32)
             _call = lambda: run_knn_pipeline_keqb(N, m, d, k, b, X, Xn, Dloc, Gdst, Gidx)
         else:
-            Gdst = torch.full((N, m, k + 1), torch.finfo(torch.float64).max, dtype=torch.float64)
+            Gdst = torch.full((N, m, k + 1), torch.finfo(torch.float32).max, dtype=torch.float32)
             Gidx = torch.full((N, m, k + 1), -1,                             dtype=torch.int32)
-            Ldst = torch.full((N, m, k + 1), torch.finfo(torch.float64).max, dtype=torch.float64)
+            Ldst = torch.full((N, m, k + 1), torch.finfo(torch.float32).max, dtype=torch.float32)
             Lidx = torch.full((N, m, k + 1), -1,                             dtype=torch.int32)
             _call = lambda: run_knn_pipeline(N, m, d, k, b, X, Xn, Dloc, Gdst, Gidx, Ldst, Lidx)
